hmship/requirements.txt
appdirs==1.4.0
cffi==1.9.1
cryptography==1.7.2
dateutils==0.6.6
Django==1.10.5
django-allauth==0.30.0
django-filter==1.0.1
django-grappelli==2.9.1
django-rest-auth==0.9.0
djangorestframework==3.5.4
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
mysqlclient==1.3.9
oauthlib==2.0.1
olefile==0.44
packaging==16.8
paypalrestsdk==1.12.0
Pillow==4.0.0
pyasn1==0.2.3
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
